+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This is a work-in-progress for the next QuPath release.
 * `isOverlayNG` should be turned on by default (https://github.com/qupath/qupath/issues/1244)
 * Labeled image instance export doesn't work as expected for z-stacks (https://github.com/qupath/qupath/issues/1267)
 * Fix PathClass singleton creation when a derived PathClass is requested (https://github.com/qupath/qupath/pull/1286)
+* 'Run for project' does not recognize when previous images have been deleted (https://github.com/qupath/qupath/issues/1291)
 
 ### Dependency updates
 * Bio-Formats 7.0.0